Remove constrain for 'max_surface_particles' value

+++ src/settings.cpp
@@ -656,7 +656,7 @@ void read_settings_xml()
 
     // Get maximum number of particles to be banked per surface
     if (check_for_node(node_ssw, "max_surface_particles")) {
-      max_surface_particles = std::stoi(get_node_value(node_ssw, "max_surface_particles"));
+      max_surface_particles = std::stoll(get_node_value(node_ssw, "max_surface_particles"));
     }
   }
